refactor: move tag settings component to pages
